fix: improve integrity guard to block corrupted snapshot before applying
- Move safeVersion regression check before SetSnapshotAsync to prevent
  applying corrupted snapshots that would be used as base for catch-up
- Update LastGood fields only when external store save succeeds to avoid
  false regression detection on next activation
- Change LastGoodEventsProcessed from int to long to prevent overflow
- Add validation for eventsProcessed > 0 with empty lastSortableUniqueId

This addresses the issue where corrupted snapshots could be applied to
the projection actor before the integrity guard detected the regression,
potentially leaving corrupted data as the base state for full catch-up.

+++ dcb/src/Sekiban.Dcb.Orleans.Core/Grains/MultiProjectionGrainState.cs
@@ -47,5 +47,5 @@ public class MultiProjectionGrainState
     public long LastGoodOriginalSizeBytes { get; set; }
 
     [Id(12)]
-    public int LastGoodEventsProcessed { get; set; }
+    public long LastGoodEventsProcessed { get; set; }
 }
